Rename README.md to README.txt
This repository is a media add-on for Kodi. Our goal is to create an environment on Kodi that acts like a traditional TV station on one hand and NetFlix on the other. The first thing to accomplish is to identify the user as a free or premium subscriber. We will store name, mailing and billing addresses, date of birth; and offer to store credit card number, name on card, card expiration date, 3 digit security code for the subscriber's convenience. The login also should have a "remember me on this device" option to keep the login process safe and easy, just like with NetFlix.
So, the environment has four icons that do the following:
1) Opens the TV stream that we are streaming from our server
2) Opens the radio stream that we are streaming from our server
3) Opens the environment that presents our video on-demand offerings 
4) Opens the environment that presents our radio podcast offerings